requirements.txt
Django==1.5.1
MySQL-python==1.2.4
South==0.8.1
distribute==0.6.31
django-tastypie==0.9.15
gunicorn==0.17.4
mimeparse==0.1.3
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
